+++ biweeklybudget/ofxgetter.py
@@ -44,8 +44,9 @@ from copy import deepcopy
 from io import StringIO
 import importlib
 
-from ofxparse import OfxParser
-from ofxclient.account import Account as OfxClientAccount
+from biweeklybudget.vendored.ofxparse import OfxParser
+from biweeklybudget.vendored.ofxclient.account \
+    import Account as OfxClientAccount
 
 from biweeklybudget.utils import Vault
 from biweeklybudget import settings