+++ biweeklybudget/vendored/ofxclient/__init__.py
@@ -1,0 +1,8 @@
+from __future__ import absolute_import
+from __future__ import unicode_literals
+from biweeklybudget.vendored.ofxclient.account import (
+    Account, BrokerageAccount, CreditCardAccount, BankAccount
+)
+from biweeklybudget.vendored.ofxclient.client import Client
+from biweeklybudget.vendored.ofxclient.institution import Institution
+from biweeklybudget.vendored.ofxclient.version import __version__